+++ .eslintrc.js
@@ -25,9 +25,5 @@ module.exports = {
       version: 'detect',
     },
   },
-  ignorePatterns: [
-    'public/client.js',
-    'public/js/iframeResizer.contentWindow.min.js',
-    'public/js/iframeResizer.min.js',
-  ],
+  ignorePatterns: ['public/client.js'],
 };